git-svn-id: http://192.168.20.11/sequoiadb/trunk@27143 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ testcase_new/story/js/FileUmask/commlib.js
@@ -43,7 +43,10 @@ FileTest.prototype.init = function()
 
 FileTest.prototype.release = function()
 {
-   this.cmd.run( "rm -rf " + this.filename ) ;    // 删除文件
+   if( this.filename != undefined )
+   {
+      this.cmd.run( "rm -rf " + this.filename ) ;    // 删除文件
+   }
    if( this.remote != undefined )
    {
       this.remote.close() ;    // 断开连接
